Document custom morpher usage
Add "Custom Morphers" document page.

The documentation assumes that we support two morphers: "morphdom" and "alpine", and that each morpher can be configured with the "MORPHER_OPTIONS" attribute.

Besides, RELOAD_SCRIPT_ELEMENTS moves from the top level to a
morpher-specific setting, and is considered an option that is only
supported by morphdom.